requirements.txt
# 项目依赖列表
flake8>=6.1.0
pytest>=8.0.0
PyYAML>=6.0
openai